An example to use the lib
In the example code, strings are formed using the returned characters and then converted to float.
'*' is used as '.'
'A' is used as + operator
'B' is used as - operator
'C' is used as CANCEL operator. It discards the last input
'D' is used for ENTER operation. When pressed, it converts string to decimal (float) and assigns to a variable.